prune FastqAPI
