+++ cg/apps/balsamic/fastq.py
@@ -150,3 +150,58 @@ class FastqHandler(BaseFastqHandler):
         for file in files:
             if os.path.isfile(file):
                 os.remove(file)
+
+    @staticmethod
+    def parse_header(line):
+        """Generates a dict with parsed lanes, flowcells and read numbers
+        Handle illumina's two different header formats
+        @see https://en.wikipedia.org/wiki/FASTQ_format
+
+        @HWUSI-EAS100R:6:73:941:1973#0/1
+
+            HWUSI-EAS100R   the unique instrument name
+            6   flowcell lane
+            73  tile number within the flowcell lane
+            941     'x'-coordinate of the cluster within the tile
+            1973    'y'-coordinate of the cluster within the tile
+            #0  index number for a multiplexed sample (0 for no indexing)
+            /1  the member of a pair, /1 or /2 (paired-end or mate-pair reads only)
+
+        Versions of the Illumina pipeline since 1.4 appear to use #NNNNNN
+        instead of #0 for the multiplex ID, where NNNNNN is the sequence of the
+        multiplex tag.
+
+        With Casava 1.8 the format of the '@' line has changed:
+
+        @EAS139:136:FC706VJ:2:2104:15343:197393 1:Y:18:ATCACG
+
+            EAS139  the unique instrument name
+            136     the run id
+            FC706VJ     the flowcell id
+            2   flowcell lane
+            2104    tile number within the flowcell lane
+            15343   'x'-coordinate of the cluster within the tile
+            197393  'y'-coordinate of the cluster within the tile
+            1   the member of a pair, 1 or 2 (paired-end or mate-pair reads only)
+            Y   Y if the read is filtered, N otherwise
+            18  0 when none of the control bits are on, otherwise it is an even number
+            ATCACG  index sequence
+        """
+
+        fastq_meta = {"lane": None, "flowcell": None, "readnumber": None}
+
+        parts = line.split(":")
+        if len(parts) == 5:  # @HWUSI-EAS100R:6:73:941:1973#0/1
+            fastq_meta["lane"] = parts[1]
+            fastq_meta["flowcell"] = "XXXXXX"
+            fastq_meta["readnumber"] = parts[-1].split("/")[-1]
+        if len(parts) == 10:  # @EAS139:136:FC706VJ:2:2104:15343:197393 1:Y:18:ATCACG
+            fastq_meta["lane"] = parts[3]
+            fastq_meta["flowcell"] = parts[2]
+            fastq_meta["readnumber"] = parts[6].split(" ")[-1]
+        if len(parts) == 7:  # @ST-E00201:173:HCLCGALXX:1:2106:22516:34834/1
+            fastq_meta["lane"] = parts[3]
+            fastq_meta["flowcell"] = parts[2]
+            fastq_meta["readnumber"] = parts[-1].split("/")[-1]
+
+        return fastq_meta
